fix: 기본 이미지 링크 Entity로 변경

+++ src/main/java/com/bookbla/americano/domain/member/controller/dto/response/MemberStyleResponse.java
@@ -13,7 +13,7 @@ public class MemberStyleResponse {
     private final String smokeType;
     private final String mbti;
     private final int height;
-    private final int profileImageTypeId;
+    private final Long profileImageTypeId;
     private final String profileImageUrl;
 
     public static MemberStyleResponse of(Member member, MemberStyle memberStyle) {
